Oppretter mindre service'r som håndterer sine ting for å unngå en eksplosjon av avhengigheter, og dermed gjøre testing/mocking enklere
